refactor into main/lib model
